+++ harfbuzz/src/hb-ot-face-table-list.hh
@@ -50,9 +50,7 @@ HB_OT_TABLE (OT, head)
 HB_OT_TABLE (OT, hhea)
 HB_OT_ACCELERATOR (OT, hmtx)
 HB_OT_TABLE (OT, OS2)
-#if !defined(HB_NO_OT_FONT_GLYPH_NAMES) || !defined(HB_NO_METRICS)
 HB_OT_ACCELERATOR (OT, post)
-#endif
 
 /* Vertical layout. */
 HB_OT_TABLE (OT, vhea)