+++ harfbuzz/src/hb-ot-font.cc
@@ -184,7 +184,6 @@ hb_ot_get_glyph_extents (hb_font_t *font,
   return ret;
 }
 
-#ifndef HB_NO_OT_FONT_GLYPH_NAMES
 hb_bool_t
 hb_ot_get_glyph_name (void *font_data,
 		      hb_codepoint_t glyph,
@@ -193,15 +192,6 @@ hb_ot_get_glyph_name (void *font_data,
   const hb_ot_face_t *ot_face = (const hb_ot_face_t *) font_data;
   return ot_face->post->get_glyph_name (glyph, name, size);
 }
-hb_bool_t
-hb_ot_get_glyph_from_name (void *font_data,
-			   const char *name, int len,
-			   hb_codepoint_t *glyph)
-{
-  const hb_ot_face_t *ot_face = (const hb_ot_face_t *) font_data;
-  return ot_face->post->get_glyph_from_name (name, len, glyph);
-}
-#endif
 
 hb_bool_t
 hb_ot_get_font_h_extents (hb_font_t *font, hb_font_extents_t *metrics)