+++ harfbuzz/src/hb-shaper.hh
@@ -36,10 +36,7 @@ typedef hb_bool_t hb_shape_func_t (hb_shape_plan_t    *shape_plan,
 				   const hb_feature_t *features,
 				   unsigned int        num_features);
 
-#define HB_SHAPER_IMPLEMENT(name) \
-	extern "C" HB_INTERNAL hb_shape_func_t _hb_##name##_shape;
-#include "hb-shaper-list.hh"
-#undef HB_SHAPER_IMPLEMENT
+extern "C" HB_INTERNAL hb_shape_func_t _hb_ot_shape;
 
 struct hb_shaper_entry_t {
   char name[16];
@@ -55,13 +52,11 @@ struct hb_shaper_lazy_loader_t;
 
 #define HB_SHAPER_ORDER(Shaper) \
   HB_PASTE (HB_SHAPER_ORDER_, Shaper)
+  
 enum hb_shaper_order_t
 {
   _HB_SHAPER_ORDER_ORDER_ZERO,
-#define HB_SHAPER_IMPLEMENT(Shaper) \
-      HB_SHAPER_ORDER (Shaper),
-#include "hb-shaper-list.hh"
-#undef HB_SHAPER_IMPLEMENT
+  HB_SHAPER_ORDER_ot,
   _HB_SHAPERS_COUNT_PLUS_ONE,
   HB_SHAPERS_COUNT = _HB_SHAPERS_COUNT_PLUS_ONE - 1,
 };
@@ -111,24 +106,15 @@ struct hb_shaper_object_dataset_t
   void init0 (Object *parent_data)
   {
     this->parent_data = parent_data;
-#define HB_SHAPER_IMPLEMENT(shaper) shaper.init0 ();
-#include "hb-shaper-list.hh"
-#undef HB_SHAPER_IMPLEMENT
+    ot.init0 ();
   }
   void fini ()
   {
-#define HB_SHAPER_IMPLEMENT(shaper) shaper.fini ();
-#include "hb-shaper-list.hh"
-#undef HB_SHAPER_IMPLEMENT
+    ot.fini ();
   }
 
   Object *parent_data; /* MUST be JUST before the lazy loaders. */
-#define HB_SHAPER_IMPLEMENT(shaper) \
-	hb_shaper_lazy_loader_t<Object, HB_SHAPER_ORDER(shaper), \
-				typename hb_shaper_object_data_type_t<HB_SHAPER_ORDER(shaper), Object>::value \
-			       > shaper;
-#include "hb-shaper-list.hh"
-#undef HB_SHAPER_IMPLEMENT
+  hb_shaper_lazy_loader_t<Object, HB_SHAPER_ORDER(ot), typename hb_shaper_object_data_type_t<HB_SHAPER_ORDER(ot), Object>::value> ot;
 };
 
 #endif /* HB_SHAPER_HH */